requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
Django==4.0.6
django-tastypie==0.14.4
djangorestframework==3.13.1
idna==3.3
lxml==4.9.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-mimeparse==1.6.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
